refactor: user.dto -> auth.dto
