+++ src/api/routes/AuthRouter.ts
@@ -1,7 +1,7 @@
 import express from "express";
 import AuthController from "../controllers/AuthController.js";
 import validate from "../middlewares/validate.js";
-import { loginSchema, signupSchema } from "../types/dtos/user.dto.js";
+import { loginSchema, signupSchema } from "../types/dtos/auth.dto.js";
 const router = express.Router();
 
 router.post("/signup", validate(signupSchema), AuthController.signup);
