+++ src/api/services/AuthService.ts
@@ -4,7 +4,7 @@ import { hashing, compareWords } from "../libs/hashing.js";
 import { generateTokens } from "../libs/token.js";
 import type { CustomError } from "src/api/types/error.js";
 import * as AuthRepository from "../repositories/AuthRepository.js";
-import type { SignupDto, LoginDto } from "../types/dtos/user.dto.js";
+import type { SignupDto, LoginDto } from "../types/dtos/auth.dto.js";
 
 const AuthService = {
   async signup(signupData: SignupDto) {
